refactor: 디렉토리 변경, 불필요 폴더 삭제, docs수정
