+++ README.md
@@ -4,15 +4,15 @@
 
 ## 기능 요구 사항 📝
 
-- [ ] 타이머 페이지를 볼 수 있다.
-- [ ] 버튼은 시작, 재설정, 타이머 수정으로 이루어져 있다.
-- [ ] 타이머 시작
-  - [ ] 시작을 누르면, 타이머가 돌아간다. -> 시작 버튼이 중지 버튼이 바뀐다.
-- [ ] 타이머 수정
-  - [ ] 타이머 수정을 누르면, 모달창이 띄워진다.
-  - [ ] 모달창의 역할은 타이머 시간을 설정할 수 있다. (모달창 : 00분 00초)
-  - [ ] 시간을 설정하면, 타이머 시간이 설정한 값으로 초기화된다.
-- [ ] 타이머 재설정
-  - [ ] 재설정을 누르면, 타이머 시간을 설정한 시간으로 초기화한다.
-- [ ] 타이머 중지
-  - [ ] 중지를 누르면, 타이머 시간이 멈춘다.
+- [x] 타이머 페이지를 볼 수 있다.
+- [x] 버튼은 시작, 재설정, 타이머 수정으로 이루어져 있다.
+- [x] 타이머 시작
+  - [x] 시작을 누르면, 타이머가 돌아간다. -> 시작 버튼이 중지 버튼이 바뀐다.
+- [x] 타이머 수정
+  - [x] 타이머 수정을 누르면, 프롬프트 창이 띄워진다.
+  - [x] 프롬프트창의 역할은 타이머 시간을 설정할 수 있다. (모달창 : 00분 00초)
+  - [x] 시간을 설정하면, 타이머 시간이 설정한 값으로 초기화된다.
+- [x] 타이머 재설정
+  - [x] 재설정을 누르면, 타이머 시간을 설정한 시간으로 초기화한다.
+- [x] 타이머 중지
+  - [x] 중지를 누르면, 타이머 시간이 멈춘다.
